TabGridViewAdded, customSheetView now cleaner
