+++ Bilbary for iPhone/CustomSheetView.swift
@@ -4,7 +4,6 @@
 //
 //  Created by SHOHJAHON on 13/07/24.
 //
-
 import SwiftUI
 import SwiftData
 
@@ -30,53 +29,12 @@ struct CustomSheetView: View {
     @State
     private var timeSpent: TimeInterval?
 
-    let columns: [GridItem] = [
-        .init(.flexible(minimum: 0, maximum: .infinity), alignment: .center),
-        .init(.flexible(minimum: 0, maximum: .infinity), alignment: .center),
-        .init(.flexible(minimum: 0, maximum: .infinity), alignment: .center),
-        .init(.flexible(minimum: 0, maximum: .infinity), alignment: .center),
-        .init(.flexible(minimum: 0, maximum: .infinity), alignment: .center)
-    ]
-
     var body: some View {
         ZStack {
             Rectangle().fill((.black.opacity(0.5)))
 
             VStack {
-                LazyVGrid(columns: columns) {
-                    TabButtonItem(
-                        title: "Book",
-                        systemImage: "book"
-                    ).gesture(
-                        dragSheetGesture(observer: coordinator)
-                    )
-                    TabButtonItem(
-                        title: "Reading Time",
-                        systemImage: "hourglass"
-                    ).gesture(
-                        dragSheetGesture(observer: coordinator)
-                    )
-                    TabButtonItem(
-                        title: "Like",
-                        systemImage: "heart"
-                    ).gesture(
-                        dragSheetGesture(observer: coordinator)
-                    )
-                    TabButtonItem(
-                        title: "Streak Goal",
-                        systemImage: "circlebadge.2"
-                    ).gesture(
-                        dragSheetGesture(observer: coordinator)
-                    )
-                    TabButtonItem(
-                        title: "Book",
-                        systemImage: "textformat.size"
-                    ).gesture(
-                        dragSheetGesture(observer: coordinator)
-                    )
-
-                }
-                .padding()
+                TabGridView()
 
                 VStack(spacing: 16) {
